Rework GCE so it doesn't apply to MacOS
Now that we have a MacOS build task, not every task uses the
gce_instance, but the defaults get applied to the MacOS task as the
top-level is applied. This errors because we didn't specify the project
or image.

Change GCE to use YAML anchors and apply these to all GCE tasks. The
anchor has to be included first in each task so it doesn't overwrite
task-specific settings; this currently only affects FreeBSD, which
changes the disk size.