MacOS: Set CPPFLAGS and LDFLAGS, use Xcode image
Homebrew doesn't link openssl by default so as not to interfere with the
system's libssl.

Set CPPFLAGS and LDFLAGS so autoconf can find libssl.

The infrastructure/BoxPlatform.pm script needs to run `xcodebuild` to
check which version of XCode is installed. This doesn't work with the
Command-Line Tools only installation.